Changed handling of boundary conditions
Now each boundary edge is marked with a negative number.

The negative number is mapped (with std::map) to a BoundaryInfo object,
that determines if the boundary is periodic or not.
Periodic boundary now supported for both explicit and implicit codes.

Instead of being built in, the boundary conditions are passed to the
constructor for each equation.